requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.2
django-rest-knox>=4.2
django-cors-headers>=4.3
gunicorn>=21.2
whitenoise>=6.6